pyproject.toml
[tool.poetry]
name = "llm-doc-assistant"
version = "0.1.0"
description = "LLM assistant with RAG for pdf document processing"
authors = ["IGPO <igpo@ukr.net>"]
packages = [{ include = "llm_doc_assistant" }]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "0.115.12"
uvicorn = "0.34.2"
langchain = "0.3.25"
openai = "1.78.1"
pypdf = "5.5.0"
annotated-types = "0.7.0"
anyio = "4.9.0"
certifi = "2025.4.26"
charset-normalizer = "3.4.2"
click = "8.2.0"
distro = "1.9.0"
faiss-cpu = "1.11.0"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
idna = "3.10"
jiter = "0.9.0"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
langchain-core = "0.3.60"
langchain-text-splitters = "0.3.8"
langsmith = "0.3.42"
numpy = "2.2.5"
orjson = "3.10.18"
packaging = "24.2"
pydantic = "2.11.4"
pydantic-core = "2.33.2"
python-dotenv = "1.1.0"
pyyaml = "6.0.2"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.41"
starlette = "0.46.2"
tenacity = "9.1.2"
tqdm = "4.67.1"
typing-inspection = "0.4.0"
typing-extensions = "4.13.2"
urllib3 = "2.4.0"
zstandard = "0.23.0"
langchain-openai = "^0.3.17"
langchain-community = "^0.3.24"
jinja2 = "^3.1.6"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
